+++ src/page/App/components/ActionEditor/ActionEditorPanel/EditorHeader/interface.ts
@@ -1,0 +1,9 @@
+export interface DropDownListProps {
+  onDeleteActionItem: () => void
+  onDuplicateActionItem: () => void
+}
+
+export interface ActionEditorHeaderMoreButtonProps extends DropDownListProps {}
+
+export interface ActionEditorHeaderProps
+  extends ActionEditorHeaderMoreButtonProps {}
