+++ src/page/App/components/ActionEditor/context.ts
@@ -3,4 +3,6 @@ import { ActionEditorContextProps } from "./interface"
 
 export const ActionEditorContext = createContext<ActionEditorContextProps>({
   editorHeight: 0,
+  baseActionApi: "",
+  isActionDirty: false,
 })
