+++ src/page/App/components/ActionEditor/index.tsx
@@ -263,6 +263,7 @@ export const ActionEditor: FC<ActionEditorProps> = (props) => {
         setActionListLoading,
         setIsActionDirty,
         baseActionApi,
+        isActionDirty,
       }}
     >
       <div className={className}>
